Refactored code, removed code smells.
