+++ openmetadata-ui/src/main/resources/ui/src/components/EntityTable/EntityTable.component.tsx
@@ -13,13 +13,18 @@
 
 import { faCaretDown, faCaretRight } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { Popover } from 'antd';
+import { Popover, Table } from 'antd';
 import classNames from 'classnames';
 import { cloneDeep, isEmpty, isNil, isUndefined, lowerCase } from 'lodash';
 import { EntityFieldThreads, EntityTags, TagOption } from 'Models';
-import React, { Fragment, useEffect, useState } from 'react';
+import React, {
+  Fragment,
+  useCallback,
+  useEffect,
+  useMemo,
+  useState,
+} from 'react';
 import { Link, useHistory } from 'react-router-dom';
-import { useExpanded, useTable } from 'react-table';
 import { useAuthContext } from '../../authentication/auth-provider/AuthProvider';
 import { FQN_SEPARATOR_CHAR } from '../../constants/char.constants';
 import { getTableDetailsPath } from '../../constants/constants';
@@ -68,8 +73,9 @@ import RichTextEditorPreviewer from '../common/rich-text-editor/RichTextEditorPr
 import { ModalWithMarkdownEditor } from '../Modals/ModalWithMarkdownEditor/ModalWithMarkdownEditor';
 import TagsContainer from '../tags-container/tags-container';
 import TagsViewer from '../tags-viewer/tags-viewer';
-import { TABLE_HEADERS } from './EntityTable.constant';
+import { TABLE_HEADERS_V1 } from './EntityTable.constants';
 import { EntityTableProps } from './EntityTable.interface';
+import './EntityTable.style.less';
 
 const EntityTable = ({
   tableColumns,
@@ -88,7 +94,6 @@ const EntityTable = ({
   const { isAdminUser, userPermissions } = useAuth();
   const { isAuthDisabled } = useAuthContext();
   const history = useHistory();
-  const columns = TABLE_HEADERS;
 
   const [searchedColumns, setSearchedColumns] = useState<ModifiedTableColumn[]>(
     []
@@ -99,21 +104,6 @@ const EntityTable = ({
     [searchedColumns]
   );
 
-  const {
-    getTableProps,
-    getTableBodyProps,
-    headerGroups,
-    rows,
-    prepareRow,
-    toggleAllRowsExpanded,
-  } = useTable(
-    {
-      columns,
-      data,
-      autoResetExpanded: false,
-    },
-    useExpanded
-  );
   const [editColumn, setEditColumn] = useState<{
     column: Column;
     index: number;
@@ -162,6 +152,7 @@ const EntityTable = ({
         } else {
           setTagFetchFailed(true);
         }
+        setIsTagLoading(false);
       })
       .catch(() => {
         setAllTags([]);
@@ -210,7 +201,7 @@ const EntityTable = ({
       const prevTags = column?.tags?.filter((tag) => {
         return newColumnTags
           .map((tag) => tag.fqn)
-          .includes(tag?.tagFQN as string);
+          .includes(tag.tagFQN as string);
       });
 
       const newTags: Array<EntityTags> = newColumnTags
@@ -304,8 +295,6 @@ const EntityTable = ({
       } else if (!isUndefined(column.children)) {
         const searchedChildren = searchInColumns(column.children, searchText);
         if (searchedChildren.length > 0) {
-          toggleAllRowsExpanded(true);
-
           return [
             ...searchedCols,
             {
@@ -329,7 +318,7 @@ const EntityTable = ({
     hasPemission(Operation.EditDescription, EntityType.TABLE, userPermissions);
   /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
   const getColumnName = (cell: any) => {
-    const fqn = cell?.row?.original?.fullyQualifiedName || '';
+    const fqn = cell?.fullyQualifiedName || '';
     const columnName = getPartialNameFromTableFQN(fqn, [FqnPart.NestedColumn]);
     // wrap it in quotes if dot is present
 
@@ -406,13 +395,12 @@ const EntityTable = ({
     handleEditColumn(column, index);
   };
 
-  /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
-  const getRequestDescriptionElement = (cell: any) => {
-    const hasDescription = Boolean(cell.value);
+  const getRequestDescriptionElement = (cell: ModifiedTableColumn) => {
+    const hasDescription = Boolean(cell.fullyQualifiedName);
 
     return (
       <button
-        className="tw-w-8 tw-h-8 tw-mr-1 tw-flex-none link-text focus:tw-outline-none tw-opacity-0 group-hover:tw-opacity-100"
+        className="tw-w-8 tw-h-8 tw-mr-1 tw-flex-none link-text focus:tw-outline-none hover-cell-icon"
         data-testid="request-description"
         onClick={() =>
           hasDescription
@@ -446,7 +434,7 @@ const EntityTable = ({
 
     return (
       <button
-        className="tw-w-8 tw-h-8 tw-mr-1 tw-flex-none link-text focus:tw-outline-none tw-opacity-0 group-hover:tw-opacity-100 tw-align-top"
+        className="tw-w-8 tw-h-8 tw-mr-1 tw-flex-none link-text focus:tw-outline-none tw-align-top hover-cell-icon"
         data-testid="request-tags"
         onClick={() =>
           hasTags ? onUpdateTagsHandler(cell) : onRequestTagsHandler(cell)
@@ -463,16 +451,457 @@ const EntityTable = ({
     );
   };
 
-  /* eslint-disable-next-line */
-  const handleTagContainerClick = (row: any) => {
-    if (!editColumnTag) {
-      handleEditColumnTag(row.original, row.id);
-      // Fetch tags and terms only once
-      if (allTags.length === 0 || tagFetchFailed) {
-        fetchTagsAndGlossaryTerms();
-      }
-    }
+  const getTestStats = (
+    record: ModifiedTableColumn | Column,
+    columnTestLength: number | undefined
+  ) => {
+    const columnTests =
+      columnTestLength && columnTestLength > 0
+        ? ((record.columnTests ?? []) as ColumnTest[])
+        : ([] as ColumnTest[]);
+
+    const failingTests = columnTests.filter((test) =>
+      test.results?.some((t) => t.testCaseStatus === TestCaseStatus.Failed)
+    );
+    const passingTests = columnTests.filter((test) =>
+      test.results?.some((t) => t.testCaseStatus === TestCaseStatus.Success)
+    );
+
+    return [failingTests, passingTests];
   };
+
+  const getColumnTestsCell = (
+    columnTestLength: number | undefined,
+    failingTests: ColumnTest[],
+    passingTests: ColumnTest[]
+  ) => {
+    return (
+      <>
+        {columnTestLength ? (
+          <>
+            {failingTests.length ? (
+              <div className="tw-flex">
+                <p className="tw-mr-2">
+                  <FontAwesomeIcon
+                    className="tw-text-status-failed"
+                    icon="times"
+                  />
+                </p>
+                <p>
+                  {`${failingTests.length}/${columnTestLength} tests failing`}
+                </p>
+              </div>
+            ) : (
+              <>
+                {passingTests.length ? (
+                  <div className="tw-flex">
+                    <div className="tw-mr-2">
+                      <FontAwesomeIcon
+                        className="tw-text-status-success"
+                        icon="check-square"
+                      />
+                    </div>
+                    <p>{`${passingTests.length} tests`}</p>
+                  </div>
+                ) : (
+                  <p>{`${columnTestLength} tests`}</p>
+                )}
+              </>
+            )}
+          </>
+        ) : (
+          '--'
+        )}
+      </>
+    );
+  };
+
+  const getDataTypeDisplayCell = (record: ModifiedTableColumn | Column) => {
+    return (
+      <>
+        {record.dataTypeDisplay ? (
+          <>
+            {isReadOnly ? (
+              <div className="tw-flex tw-flex-wrap tw-w-60 tw-overflow-x-auto">
+                <RichTextEditorPreviewer
+                  markdown={record.dataTypeDisplay.toLowerCase()}
+                />
+              </div>
+            ) : (
+              <>
+                {record.dataTypeDisplay.length > 25 ? (
+                  <span>
+                    <PopOver
+                      html={
+                        <div className="tw-break-words">
+                          <span>{record.dataTypeDisplay.toLowerCase()}</span>
+                        </div>
+                      }
+                      key="pop-over"
+                      position="bottom"
+                      theme="light"
+                      trigger="click">
+                      <div className="tw-cursor-pointer tw-underline tw-inline-block">
+                        <RichTextEditorPreviewer
+                          markdown={`${record.dataTypeDisplay
+                            .slice(0, 20)
+                            .toLowerCase()}...`}
+                        />
+                      </div>
+                    </PopOver>
+                  </span>
+                ) : (
+                  record.dataTypeDisplay.toLowerCase()
+                )}
+              </>
+            )}
+          </>
+        ) : (
+          '--'
+        )}
+      </>
+    );
+  };
+
+  const getDescriptionCell = (
+    index: number,
+    record: ModifiedTableColumn | Column
+  ) => {
+    return (
+      <div className="hover-icon-group">
+        <div className="tw-inline-block">
+          <div
+            className="tw-flex"
+            data-testid="description"
+            id={`column-description-${index}`}>
+            <div>
+              {record?.description ? (
+                <RichTextEditorPreviewer markdown={record?.description} />
+              ) : (
+                <span className="tw-no-description">No description</span>
+              )}
+            </div>
+            <div className="tw-flex tw--mt-2">
+              {!isReadOnly ? (
+                <Fragment>
+                  {checkPermission() && (
+                    <>
+                      <button
+                        className="tw-self-start tw-w-8 tw-h-8 tw-ml-1 focus:tw-outline-none tw-flex-none hover-cell-icon"
+                        onClick={() => handleUpdate(record, index)}>
+                        <SVGIcons
+                          alt="edit"
+                          icon="icon-edit"
+                          title="Edit"
+                          width="14px"
+                        />
+                      </button>
+                    </>
+                  )}
+                  {getRequestDescriptionElement(record)}
+                  {getFieldThreadElement(
+                    getColumnName(record),
+                    EntityField.DESCRIPTION,
+                    entityFieldThreads as EntityFieldThreads[],
+                    onThreadLinkSelect,
+                    EntityType.TABLE,
+                    entityFqn,
+                    `columns${ENTITY_LINK_SEPARATOR}${getColumnName(
+                      record
+                    )}${ENTITY_LINK_SEPARATOR}description`,
+                    Boolean(record)
+                  )}
+                  {getFieldThreadElement(
+                    getColumnName(record),
+                    EntityField.DESCRIPTION,
+                    entityFieldTasks as EntityFieldThreads[],
+                    onThreadLinkSelect,
+                    EntityType.TABLE,
+                    entityFqn,
+                    `columns${ENTITY_LINK_SEPARATOR}${getColumnName(
+                      record
+                    )}${ENTITY_LINK_SEPARATOR}description`,
+                    Boolean(record),
+                    ThreadType.Task
+                  )}
+                </Fragment>
+              ) : null}
+            </div>
+          </div>
+        </div>
+        {checkIfJoinsAvailable(record?.name) && (
+          <div className="tw-mt-3" data-testid="frequently-joined-columns">
+            <span className="tw-text-grey-muted tw-mr-1">
+              Frequently joined columns:
+            </span>
+            <span>
+              {getFrequentlyJoinedWithColumns(record?.name)
+                .slice(0, 3)
+                .map((columnJoin, index) => (
+                  <Fragment key={index}>
+                    {index > 0 && <span className="tw-mr-1">,</span>}
+                    <Link
+                      className="link-text"
+                      to={getTableDetailsPath(
+                        getTableFQNFromColumnFQN(
+                          columnJoin.fullyQualifiedName as string
+                        ),
+                        getPartialNameFromTableFQN(
+                          columnJoin.fullyQualifiedName as string,
+                          [FqnPart.Column]
+                        )
+                      )}>
+                      {getPartialNameFromTableFQN(
+                        columnJoin.fullyQualifiedName as string,
+                        [FqnPart.Database, FqnPart.Table, FqnPart.Column],
+                        FQN_SEPARATOR_CHAR
+                      )}
+                    </Link>
+                  </Fragment>
+                ))}
+
+              {getFrequentlyJoinedWithColumns(record?.name).length > 3 && (
+                <PopOver
+                  html={
+                    <div className="tw-text-left">
+                      {getFrequentlyJoinedWithColumns(record?.name)
+                        ?.slice(3)
+                        .map((columnJoin, index) => (
+                          <Fragment key={index}>
+                            <a
+                              className="link-text tw-block tw-py-1"
+                              href={getTableDetailsPath(
+                                getTableFQNFromColumnFQN(
+                                  columnJoin?.fullyQualifiedName
+                                ),
+                                getPartialNameFromTableFQN(
+                                  columnJoin?.fullyQualifiedName,
+                                  [FqnPart.Column]
+                                )
+                              )}>
+                              {getPartialNameFromTableFQN(
+                                columnJoin?.fullyQualifiedName,
+                                [
+                                  FqnPart.Database,
+                                  FqnPart.Table,
+                                  FqnPart.Column,
+                                ]
+                              )}
+                            </a>
+                          </Fragment>
+                        ))}
+                    </div>
+                  }
+                  position="bottom"
+                  theme="light"
+                  trigger="click">
+                  <span className="show-more tw-ml-1 tw-underline">...</span>
+                </PopOver>
+              )}
+            </span>
+          </div>
+        )}
+      </div>
+    );
+  };
+
+  const getTagsCell = (index: number, record: ModifiedTableColumn | Column) => {
+    return (
+      <div className="hover-icon-group">
+        {isReadOnly ? (
+          <div className="tw-flex tw-flex-wrap">
+            <TagsViewer sizeCap={-1} tags={record?.tags || []} />
+          </div>
+        ) : (
+          <div
+            className={classNames(
+              `tw-flex tw-justify-content`,
+              editColumnTag?.index === index || !isEmpty(record.tags)
+                ? 'tw-flex-col tw-items-start'
+                : 'tw-items-center'
+            )}
+            data-testid="tags-wrapper"
+            onClick={() => {
+              if (!editColumnTag) {
+                handleEditColumnTag(record, index);
+                // Fetch tags and terms only once
+                if (allTags.length === 0 || tagFetchFailed) {
+                  fetchTagsAndGlossaryTerms();
+                }
+              }
+            }}>
+            <NonAdminAction
+              html={getHtmlForNonAdminAction(Boolean(owner))}
+              isOwner={hasEditAccess}
+              permission={Operation.EditTags}
+              position="left"
+              trigger="click">
+              <TagsContainer
+                showAddTagButton
+                editable={editColumnTag?.index === index}
+                isLoading={isTagLoading && editColumnTag?.index === index}
+                selectedTags={record?.tags || []}
+                size="small"
+                tagList={allTags}
+                type="label"
+                onCancel={() => {
+                  handleTagSelection();
+                }}
+                onSelectionChange={(tags) => {
+                  handleTagSelection(tags, record?.name);
+                }}
+              />
+            </NonAdminAction>
+            <div className="tw-mt-1 tw-flex">
+              {getRequestTagsElement(record)}
+              {getFieldThreadElement(
+                getColumnName(record),
+                'tags',
+                entityFieldThreads as EntityFieldThreads[],
+                onThreadLinkSelect,
+                EntityType.TABLE,
+                entityFqn,
+                `columns${ENTITY_LINK_SEPARATOR}${getColumnName(
+                  record
+                )}${ENTITY_LINK_SEPARATOR}tags`,
+                Boolean(record?.name?.length)
+              )}
+              {getFieldThreadElement(
+                getColumnName(record),
+                EntityField.TAGS,
+                entityFieldTasks as EntityFieldThreads[],
+                onThreadLinkSelect,
+                EntityType.TABLE,
+                entityFqn,
+                `${EntityField.COLUMNS}${ENTITY_LINK_SEPARATOR}${getColumnName(
+                  record
+                )}${ENTITY_LINK_SEPARATOR}${EntityField.TAGS}`,
+                Boolean(record?.name),
+                ThreadType.Task
+              )}
+            </div>
+          </div>
+        )}
+      </div>
+    );
+  };
+
+  const renderCell = useCallback(
+    (key: string, record: ModifiedTableColumn | Column, index: number) => {
+      const columnTestLength = record?.columnTests?.length;
+
+      const [failingTests, passingTests] = getTestStats(
+        record,
+        columnTestLength
+      );
+
+      switch (key) {
+        case TABLE_HEADERS_V1.columnTests:
+          return getColumnTestsCell(
+            columnTestLength,
+            failingTests,
+            passingTests
+          );
+
+        case TABLE_HEADERS_V1.dataTypeDisplay:
+          return getDataTypeDisplayCell(record);
+
+        case TABLE_HEADERS_V1.description:
+          return getDescriptionCell(index, record);
+
+        case TABLE_HEADERS_V1.tags:
+          return getTagsCell(index, record);
+
+        default:
+          return (
+            <Fragment>
+              {isReadOnly ? (
+                <div className="tw-inline-block">
+                  <RichTextEditorPreviewer markdown={record.name} />
+                </div>
+              ) : (
+                <span>
+                  {prepareConstraintIcon(record.name, record.constraint)}
+                  <span className="tw-ml-4">{record.name}</span>
+                </span>
+              )}
+            </Fragment>
+          );
+      }
+    },
+    [editColumnTag, isTagLoading, handleUpdate, handleTagSelection]
+  );
+
+  const columns = useMemo(
+    () => [
+      {
+        title: 'Name',
+        dataIndex: 'name',
+        key: 'name',
+        accessor: 'name',
+        ellipsis: true,
+        width: 180,
+        render: (
+          _: Array<unknown>,
+          record: ModifiedTableColumn,
+          index: number
+        ) => renderCell(TABLE_HEADERS_V1.name, record, index),
+      },
+      {
+        title: 'Type',
+        dataIndex: 'dataTypeDisplay',
+        key: 'dataTypeDisplay',
+        accessor: 'dataTypeDisplay',
+        ellipsis: true,
+        width: 200,
+        render: (
+          _: Array<unknown>,
+          record: ModifiedTableColumn,
+          index: number
+        ) => {
+          return renderCell(TABLE_HEADERS_V1.dataTypeDisplay, record, index);
+        },
+      },
+      {
+        title: 'Data Quality',
+        dataIndex: 'columnTests',
+        key: 'columnTests',
+        accessor: 'columnTests',
+        width: 200,
+        render: (
+          _: Array<unknown>,
+          record: ModifiedTableColumn,
+          index: number
+        ) => {
+          return renderCell(TABLE_HEADERS_V1.columnTests, record, index);
+        },
+      },
+      {
+        title: 'Description',
+        dataIndex: 'description',
+        key: 'description',
+        accessor: 'description',
+        render: (
+          _: Array<unknown>,
+          record: ModifiedTableColumn,
+          index: number
+        ) => renderCell(TABLE_HEADERS_V1.description, record, index),
+      },
+      {
+        title: 'Tags',
+        dataIndex: 'tags',
+        key: 'tags',
+        accessor: 'tags',
+        width: 272,
+        render: (
+          _: Array<unknown>,
+          record: ModifiedTableColumn | Column,
+          index: number
+        ) => renderCell(TABLE_HEADERS_V1.tags, record, index),
+      },
+    ],
+    [editColumnTag, isTagLoading, renderCell]
+  );
 
   useEffect(() => {
     if (!searchText) {
@@ -483,435 +912,31 @@ const EntityTable = ({
     }
   }, [searchText, tableColumns]);
 
-  useEffect(() => {
-    toggleAllRowsExpanded(isReadOnly);
-  }, []);
-
   return (
-    <div className="tw-table-responsive" id="schemaTable">
-      <table
-        className="tw-w-full"
-        {...getTableProps()}
-        data-testid="entity-table">
-        <thead data-testid="table-header">
-          {/* eslint-disable-next-line */}
-          {headerGroups.map((headerGroup: any, index: number) => (
-            <tr
-              className="tableHead-row"
-              key={index}
-              {...headerGroup.getHeaderGroupProps()}>
-              {/* eslint-disable-next-line */}
-              {headerGroup.headers.map((column: any, index: number) => (
-                <th
-                  className={classNames('tableHead-cell', {
-                    'tw-w-60':
-                      column.id === 'tags' || column.id === 'columnTests',
-                  })}
-                  data-testid={column.id}
-                  key={index}
-                  {...column.getHeaderProps()}>
-                  {column.render('Header')}
-                </th>
-              ))}
-            </tr>
-          ))}
-        </thead>
-
-        <tbody {...getTableBodyProps()} data-testid="table-body">
-          {/* eslint-disable-next-line */}
-          {rows.map((row: any, index: number) => {
-            prepareRow(row);
-
-            return (
-              <tr
-                className={classNames('tableBody-row')}
-                data-testid="row"
-                key={index}
-                {...row.getRowProps()}>
-                {/* eslint-disable-next-line */}
-                {row.cells.map((cell: any, index: number) => {
-                  const columnTests =
-                    cell.column.id === 'columnTests'
-                      ? ((cell.value ?? []) as ColumnTest[])
-                      : ([] as ColumnTest[]);
-                  const columnTestLength = columnTests.length;
-                  const failingTests = columnTests.filter((test) =>
-                    test.results?.some(
-                      (t) => t.testCaseStatus === TestCaseStatus.Failed
-                    )
-                  );
-                  const passingTests = columnTests.filter((test) =>
-                    test.results?.some(
-                      (t) => t.testCaseStatus === TestCaseStatus.Success
-                    )
-                  );
-
-                  return (
-                    <td
-                      className={classNames(
-                        'tableBody-cell tw-group tw-relative tw-align-baseline'
-                      )}
-                      key={index}
-                      {...cell.getCellProps()}>
-                      {row.canExpand && cell.column.id === 'name' ? (
-                        <span
-                          {...row.getToggleRowExpandedProps({})}
-                          className="tw-mr-2 tw-cursor-pointer"
-                          style={{
-                            marginLeft: `${row.depth * 35}px`,
-                          }}>
-                          <FontAwesomeIcon
-                            icon={row.isExpanded ? faCaretDown : faCaretRight}
-                          />
-                        </span>
-                      ) : null}
-
-                      {cell.column.id === 'columnTests' && (
-                        <Fragment>
-                          {columnTestLength ? (
-                            <Fragment>
-                              {failingTests.length ? (
-                                <div className="tw-flex">
-                                  <p className="tw-mr-2">
-                                    <FontAwesomeIcon
-                                      className="tw-text-status-failed"
-                                      icon="times"
-                                    />
-                                  </p>
-                                  <p>
-                                    {`${failingTests.length}/${columnTestLength} tests failing`}
-                                  </p>
-                                </div>
-                              ) : (
-                                <Fragment>
-                                  {passingTests.length ? (
-                                    <div className="tw-flex">
-                                      <div className="tw-mr-2">
-                                        <FontAwesomeIcon
-                                          className="tw-text-status-success"
-                                          icon="check-square"
-                                        />
-                                      </div>
-                                      <p>{`${passingTests.length} tests`}</p>
-                                    </div>
-                                  ) : (
-                                    <p>{`${columnTestLength} tests`}</p>
-                                  )}
-                                </Fragment>
-                              )}
-                            </Fragment>
-                          ) : (
-                            '--'
-                          )}
-                        </Fragment>
-                      )}
-
-                      {cell.column.id === 'dataTypeDisplay' && (
-                        <>
-                          {cell.value ? (
-                            <>
-                              {isReadOnly ? (
-                                <div className="tw-flex tw-flex-wrap tw-w-60 tw-overflow-x-auto">
-                                  <RichTextEditorPreviewer
-                                    markdown={cell.value.toLowerCase()}
-                                  />
-                                </div>
-                              ) : (
-                                <>
-                                  {cell.value.length > 25 ? (
-                                    <span>
-                                      <PopOver
-                                        html={
-                                          <div className="tw-break-words">
-                                            <span>
-                                              {cell.value.toLowerCase()}
-                                            </span>
-                                          </div>
-                                        }
-                                        position="bottom"
-                                        theme="light"
-                                        trigger="click">
-                                        <div className="tw-cursor-pointer tw-underline tw-inline-block">
-                                          <RichTextEditorPreviewer
-                                            markdown={`${cell.value
-                                              .slice(0, 20)
-                                              .toLowerCase()}...`}
-                                          />
-                                        </div>
-                                      </PopOver>
-                                    </span>
-                                  ) : (
-                                    cell.value.toLowerCase()
-                                  )}
-                                </>
-                              )}
-                            </>
-                          ) : (
-                            '--'
-                          )}
-                        </>
-                      )}
-
-                      {cell.column.id === 'tags' && (
-                        <>
-                          {isReadOnly ? (
-                            <div className="tw-flex tw-flex-wrap">
-                              <TagsViewer
-                                sizeCap={-1}
-                                tags={cell.value || []}
-                              />
-                            </div>
-                          ) : (
-                            <div
-                              data-testid="tags-wrapper"
-                              onClick={() => handleTagContainerClick(row)}>
-                              <NonAdminAction
-                                html={getHtmlForNonAdminAction(Boolean(owner))}
-                                isOwner={hasEditAccess}
-                                permission={Operation.EditTags}
-                                position="left"
-                                trigger="click">
-                                <TagsContainer
-                                  showAddTagButton
-                                  editable={editColumnTag?.index === row.id}
-                                  isLoading={
-                                    isTagLoading &&
-                                    editColumnTag?.index === row.id
-                                  }
-                                  selectedTags={cell.value || []}
-                                  size="small"
-                                  tagList={allTags}
-                                  type="label"
-                                  onCancel={() => {
-                                    handleTagSelection();
-                                  }}
-                                  onSelectionChange={(tags) => {
-                                    handleTagSelection(tags, row.original.name);
-                                  }}
-                                />
-                              </NonAdminAction>
-                              <div className="tw-mt-1">
-                                {getRequestTagsElement(cell)}
-                                {getFieldThreadElement(
-                                  getColumnName(cell),
-                                  'tags',
-                                  entityFieldThreads as EntityFieldThreads[],
-                                  onThreadLinkSelect,
-                                  EntityType.TABLE,
-                                  entityFqn,
-                                  `columns${ENTITY_LINK_SEPARATOR}${getColumnName(
-                                    cell
-                                  )}${ENTITY_LINK_SEPARATOR}tags`,
-                                  Boolean(cell.value.length)
-                                )}
-                                {getFieldThreadElement(
-                                  getColumnName(cell),
-                                  EntityField.TAGS,
-                                  entityFieldTasks as EntityFieldThreads[],
-                                  onThreadLinkSelect,
-                                  EntityType.TABLE,
-                                  entityFqn,
-                                  `${
-                                    EntityField.COLUMNS
-                                  }${ENTITY_LINK_SEPARATOR}${getColumnName(
-                                    cell
-                                  )}${ENTITY_LINK_SEPARATOR}${
-                                    EntityField.TAGS
-                                  }`,
-                                  Boolean(cell.value),
-                                  ThreadType.Task
-                                )}
-                              </div>
-                            </div>
-                          )}
-                        </>
-                      )}
-                      {cell.column.id === 'description' && (
-                        <div>
-                          <div className="tw-inline-block">
-                            <div
-                              className="tw-flex"
-                              data-testid="description"
-                              id={`column-description-${index}`}>
-                              <div>
-                                {cell.value ? (
-                                  <RichTextEditorPreviewer
-                                    markdown={cell.value}
-                                  />
-                                ) : (
-                                  <span className="tw-no-description">
-                                    No description{' '}
-                                  </span>
-                                )}
-                              </div>
-                              <div className="tw-flex tw--mt-2">
-                                {!isReadOnly ? (
-                                  <Fragment>
-                                    {checkPermission() && (
-                                      <button
-                                        className="tw-self-start tw-w-8 tw-h-8 tw-opacity-0 tw-ml-1 group-hover:tw-opacity-100 focus:tw-outline-none tw-flex-none"
-                                        onClick={() =>
-                                          handleUpdate(row.original, row.id)
-                                        }>
-                                        <SVGIcons
-                                          alt="edit"
-                                          icon="icon-edit"
-                                          title="Edit"
-                                          width="14px"
-                                        />
-                                      </button>
-                                    )}
-                                    {getRequestDescriptionElement(cell)}
-                                    {getFieldThreadElement(
-                                      getColumnName(cell),
-                                      EntityField.DESCRIPTION,
-                                      entityFieldThreads as EntityFieldThreads[],
-                                      onThreadLinkSelect,
-                                      EntityType.TABLE,
-                                      entityFqn,
-                                      `columns${ENTITY_LINK_SEPARATOR}${getColumnName(
-                                        cell
-                                      )}${ENTITY_LINK_SEPARATOR}description`,
-                                      Boolean(cell.value)
-                                    )}
-                                    {getFieldThreadElement(
-                                      getColumnName(cell),
-                                      EntityField.DESCRIPTION,
-                                      entityFieldTasks as EntityFieldThreads[],
-                                      onThreadLinkSelect,
-                                      EntityType.TABLE,
-                                      entityFqn,
-                                      `columns${ENTITY_LINK_SEPARATOR}${getColumnName(
-                                        cell
-                                      )}${ENTITY_LINK_SEPARATOR}description`,
-                                      Boolean(cell.value),
-                                      ThreadType.Task
-                                    )}
-                                  </Fragment>
-                                ) : null}
-                              </div>
-                            </div>
-                          </div>
-                          {checkIfJoinsAvailable(row.original.name) && (
-                            <div
-                              className="tw-mt-3"
-                              data-testid="frequently-joined-columns">
-                              <span className="tw-text-grey-muted tw-mr-1">
-                                Frequently joined columns:
-                              </span>
-                              <span>
-                                {getFrequentlyJoinedWithColumns(
-                                  row.original.name
-                                )
-                                  .slice(0, 3)
-                                  .map((columnJoin, index) => (
-                                    <Fragment key={index}>
-                                      {index > 0 && (
-                                        <span className="tw-mr-1">,</span>
-                                      )}
-                                      <Link
-                                        className="link-text"
-                                        to={getTableDetailsPath(
-                                          getTableFQNFromColumnFQN(
-                                            columnJoin?.fullyQualifiedName as string
-                                          ),
-                                          getPartialNameFromTableFQN(
-                                            columnJoin?.fullyQualifiedName as string,
-                                            [FqnPart.Column]
-                                          )
-                                        )}>
-                                        {getPartialNameFromTableFQN(
-                                          columnJoin?.fullyQualifiedName as string,
-                                          [
-                                            FqnPart.Database,
-                                            FqnPart.Table,
-                                            FqnPart.Column,
-                                          ],
-                                          FQN_SEPARATOR_CHAR
-                                        )}
-                                      </Link>
-                                    </Fragment>
-                                  ))}
-
-                                {getFrequentlyJoinedWithColumns(
-                                  row.original.name
-                                ).length > 3 && (
-                                  <PopOver
-                                    html={
-                                      <div className="tw-text-left">
-                                        {getFrequentlyJoinedWithColumns(
-                                          row.original.name
-                                        )
-                                          ?.slice(3)
-                                          .map((columnJoin, index) => (
-                                            <Fragment key={index}>
-                                              <a
-                                                className="link-text tw-block tw-py-1"
-                                                href={getTableDetailsPath(
-                                                  getTableFQNFromColumnFQN(
-                                                    columnJoin?.fullyQualifiedName as string
-                                                  ),
-                                                  getPartialNameFromTableFQN(
-                                                    columnJoin?.fullyQualifiedName as string,
-                                                    [FqnPart.Column]
-                                                  )
-                                                )}>
-                                                {getPartialNameFromTableFQN(
-                                                  columnJoin?.fullyQualifiedName as string,
-                                                  [
-                                                    FqnPart.Database,
-                                                    FqnPart.Table,
-                                                    FqnPart.Column,
-                                                  ]
-                                                )}
-                                              </a>
-                                            </Fragment>
-                                          ))}
-                                      </div>
-                                    }
-                                    position="bottom"
-                                    theme="light"
-                                    trigger="click">
-                                    <span className="show-more tw-ml-1 tw-underline">
-                                      ...
-                                    </span>
-                                  </PopOver>
-                                )}
-                              </span>
-                            </div>
-                          )}
-                        </div>
-                      )}
-                      {cell.column.id === 'name' && (
-                        <Fragment>
-                          {isReadOnly ? (
-                            <div className="tw-inline-block">
-                              <RichTextEditorPreviewer markdown={cell.value} />
-                            </div>
-                          ) : (
-                            <span
-                              style={{
-                                paddingLeft: `${
-                                  row.canExpand ? '0px' : `${row.depth * 35}px`
-                                }`,
-                              }}>
-                              {prepareConstraintIcon(
-                                cell.value,
-                                row.original.constraint
-                              )}
-                              {cell.render('Cell')}
-                            </span>
-                          )}
-                        </Fragment>
-                      )}
-                    </td>
-                  );
-                })}
-              </tr>
-            );
-          })}
-        </tbody>
-      </table>
+    <>
+      <Table
+        columns={columns}
+        data-testid="entity-table"
+        dataSource={data}
+        expandable={{
+          defaultExpandedRowKeys: [],
+          expandIcon: ({ expanded, onExpand, record }) =>
+            record.children ? (
+              <FontAwesomeIcon
+                className="tw-mr-2 tw-cursor-pointer"
+                icon={expanded ? faCaretDown : faCaretRight}
+                onClick={(e) =>
+                  onExpand(
+                    record,
+                    e as unknown as React.MouseEvent<HTMLElement, MouseEvent>
+                  )
+                }
+              />
+            ) : null,
+        }}
+        pagination={false}
+        size="small"
+      />
       {editColumn && (
         <ModalWithMarkdownEditor
           header={`Edit column: "${editColumn.column.name}"`}
@@ -919,9 +944,10 @@ const EntityTable = ({
           value={editColumn.column.description as string}
           onCancel={closeEditColumnModal}
           onSave={handleEditColumnChange}
+          // expandable={}
         />
       )}
-    </div>
+    </>
   );
 };
