+++ openmetadata-ui/src/main/resources/ui/src/components/EntityTable/EntityTable.style.less
