+++ openmetadata-ui/src/main/resources/ui/src/components/SchemaTab/SchemaTab.test.tsx
@@ -75,7 +75,7 @@ jest.mock('../SampleDataTable/SampleDataTable.component', () => {
   return jest.fn().mockReturnValue(<p>SampleDataTable</p>);
 });
 
-jest.mock('../EntityTable/EntityTableV1.component', () => {
+jest.mock('../EntityTable/EntityTable.component', () => {
   return jest.fn().mockReturnValue(<p>EntityTableV1</p>);
 });
